Move worker to jobs package
